Refactor: Rename inspection classes and split visitor logic
