+++ src/main/kotlin/org/domaframework/doma/intellij/inspection/dao/inspector/DaoMethodVariableInspection.kt
@@ -13,32 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.domaframework.doma.intellij.inspection.sql.inspector
+package org.domaframework.doma.intellij.inspection.dao.inspector
 
 import com.intellij.codeHighlighting.HighlightDisplayLevel
-import com.intellij.codeInspection.LocalInspectionTool
+import com.intellij.codeInspection.AbstractBaseJavaLocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
-import org.domaframework.doma.intellij.inspection.sql.visitor.SqlInspectionVisitor
-import org.domaframework.doma.intellij.psi.SqlVisitor
+import com.intellij.psi.PsiElementVisitor
+import org.domaframework.doma.intellij.inspection.dao.visitor.DaoMethodVariableInspectionVisitor
 
 /**
- * Code inspection for SQL bind variables
+ * Check if DAO method arguments are used in the corresponding SQL file
  */
-class SqlBindVariableValidInspector : LocalInspectionTool() {
-    override fun getDisplayName(): String = "Check where SQL bind variables are defined."
+class DaoMethodVariableInspection : AbstractBaseJavaLocalInspectionTool() {
+    override fun getDisplayName(): String = "Check usage of DAO method arguments in corresponding SQL file."
 
-    override fun getShortName(): String = "org.domaframework.doma.intellij.validBindVariable"
+    override fun getShortName(): String = "org.domaframework.doma.intellij.variablechecker"
 
     override fun getGroupDisplayName(): String = "DomaTools"
 
     override fun isEnabledByDefault(): Boolean = true
 
-    override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.Companion.ERROR
-
-    override fun runForWholeFile(): Boolean = true
+    override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.ERROR
 
     override fun buildVisitor(
         holder: ProblemsHolder,
         isOnTheFly: Boolean,
-    ): SqlVisitor = SqlInspectionVisitor(holder, this.shortName)
+    ): PsiElementVisitor = DaoMethodVariableInspectionVisitor(holder)
 }
