+++ src/main/kotlin/org/domaframework/doma/intellij/inspection/dao/inspector/SqlFileExistInspection.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright Doma Tools Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.domaframework.doma.intellij.inspection.dao.inspector
+
+import com.intellij.codeHighlighting.HighlightDisplayLevel
+import com.intellij.codeInspection.AbstractBaseJavaLocalInspectionTool
+import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.psi.PsiElementVisitor
+import org.domaframework.doma.intellij.inspection.dao.visitor.SqlFileExistInspectionVisitor
+
+/**
+ * Check for existence of SQL file
+ */
+class SqlFileExistInspection : AbstractBaseJavaLocalInspectionTool() {
+    override fun getDisplayName(): String = "Ensure the existence of SQL files for DAO methods."
+
+    override fun getShortName(): String = "org.domaframework.doma.intellij.existsqlchecker"
+
+    override fun getGroupDisplayName(): String = "DomaTools"
+
+    override fun isEnabledByDefault(): Boolean = true
+
+    override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.ERROR
+
+    override fun buildVisitor(
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+    ): PsiElementVisitor = SqlFileExistInspectionVisitor(holder)
+}
