+++ src/main/kotlin/org/domaframework/doma/intellij/inspection/dao/visitor/SqlFileExistInspectionVisitor.kt
@@ -13,14 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.domaframework.doma.intellij.inspection.dao.inspector
+package org.domaframework.doma.intellij.inspection.dao.visitor
 
-import com.intellij.codeHighlighting.HighlightDisplayLevel
-import com.intellij.codeInspection.AbstractBaseJavaLocalInspectionTool
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.JavaElementVisitor
-import com.intellij.psi.PsiElementVisitor
 import com.intellij.psi.PsiMethod
 import org.domaframework.doma.intellij.bundle.MessageBundle
 import org.domaframework.doma.intellij.common.dao.getDaoClass
@@ -28,40 +25,22 @@ import org.domaframework.doma.intellij.common.isJavaOrKotlinFileType
 import org.domaframework.doma.intellij.common.psi.PsiDaoMethod
 import org.domaframework.doma.intellij.inspection.dao.quickfix.GenerateSQLFileQuickFixFactory
 
-/**
- * Check for existence of SQL file
- */
-class SqlFileExistInspector : AbstractBaseJavaLocalInspectionTool() {
-    override fun getDisplayName(): String = "Ensure the existence of SQL files for DAO methods."
+class SqlFileExistInspectionVisitor(
+    private val holder: ProblemsHolder,
+) : JavaElementVisitor() {
+    // TODO Support Kotlin Project
+    override fun visitMethod(method: PsiMethod) {
+        super.visitMethod(method)
+        val file = method.containingFile
+        if (!isJavaOrKotlinFileType(file) || getDaoClass(file) == null) return
 
-    override fun getShortName(): String = "org.domaframework.doma.intellij.existsqlchecker"
-
-    override fun getGroupDisplayName(): String = "DomaTools"
-
-    override fun isEnabledByDefault(): Boolean = true
-
-    override fun getDefaultLevel(): HighlightDisplayLevel = HighlightDisplayLevel.ERROR
-
-    override fun buildVisitor(
-        holder: ProblemsHolder,
-        isOnTheFly: Boolean,
-    ): PsiElementVisitor {
-        // TODO Support Kotlin Project
-        return object : JavaElementVisitor() {
-            override fun visitMethod(method: PsiMethod) {
-                super.visitMethod(method)
-                val file = method.containingFile
-                if (!isJavaOrKotlinFileType(file) || getDaoClass(file) == null) return
-
-                val psiDaoMethod = PsiDaoMethod(method.project, method)
-                if (psiDaoMethod.isUseSqlFileMethod()) {
-                    checkDaoMethod(psiDaoMethod, holder)
-                }
-            }
+        val psiDaoMethod = PsiDaoMethod(method.project, method)
+        if (psiDaoMethod.isUseSqlFileMethod()) {
+            checkDaoMethod(psiDaoMethod, holder)
         }
     }
 
-    fun checkDaoMethod(
+    private fun checkDaoMethod(
         psiDaoMethod: PsiDaoMethod,
         problemHolder: ProblemsHolder,
     ) {
