+++ src/main/kotlin/org/domaframework/doma/intellij/inspection/sql/provider/SqlBindVariableProvider.kt
@@ -17,11 +17,11 @@ package org.domaframework.doma.intellij.inspection.sql.provider
 
 import com.intellij.codeInspection.InspectionToolProvider
 import com.intellij.codeInspection.LocalInspectionTool
-import org.domaframework.doma.intellij.inspection.sql.inspector.SqlBindVariableValidInspector
+import org.domaframework.doma.intellij.inspection.sql.inspector.SqlBindVariableInspection
 
-class SqlBindVariableValidProvider : InspectionToolProvider {
+class SqlBindVariableProvider : InspectionToolProvider {
     override fun getInspectionClasses(): Array<Class<out LocalInspectionTool>> =
         arrayOf(
-            SqlBindVariableValidInspector::class.java,
+            SqlBindVariableInspection::class.java,
         )
 }
