+++ src/test/kotlin/org/domaframework/doma/intellij/inspection/sql/ParameterDefinedTest.kt
@@ -16,9 +16,9 @@
 package org.domaframework.doma.intellij.inspection.sql
 
 import org.domaframework.doma.intellij.DomaSqlTest
-import org.domaframework.doma.intellij.inspection.sql.inspector.SqlBindVariableValidInspector
-import org.domaframework.doma.intellij.inspection.sql.inspector.SqlFunctionCallInspector
-import org.domaframework.doma.intellij.inspection.sql.inspector.SqlLoopDirectiveTypeInspector
+import org.domaframework.doma.intellij.inspection.sql.inspector.SqlBindVariableInspection
+import org.domaframework.doma.intellij.inspection.sql.inspector.SqlFunctionCallInspection
+import org.domaframework.doma.intellij.inspection.sql.inspector.SqlLoopDirectiveTypeInspection
 
 /**
  * A test that inspects whether a bind variable's parameters are defined.
@@ -46,9 +46,9 @@ class ParameterDefinedTest : DomaSqlTest() {
             "$testDaoName/emptyImplementCustomFunctions.sql",
         )
         myFixture.enableInspections(
-            SqlBindVariableValidInspector(),
-            SqlLoopDirectiveTypeInspector(),
-            SqlFunctionCallInspector(),
+            SqlBindVariableInspection(),
+            SqlLoopDirectiveTypeInspection(),
+            SqlFunctionCallInspection(),
         )
     }
